Add `g:Prosession_ignore_expr`. Fix #80
This is a `Funcref` that must return a boolean value, either `v:true` or
`v:false`, upon invocation to determine if a directory should be ignored
by prosession.

```viml
  let g:Prosession_ignore_expr = {->v:false}
```

This could be used to ignore directories by using some specific
criteria. Following is an example to ignore all directories that
are not git repos.

```viml
  let g:Prosession_ignore_expr = {-> !isdirectory('.git')}
```